Adds historical blacklisting.
Once an actor is blacklisted and goes away for the expire period, their
records are removed. It is nice to be able to keep a list of actors who
have been blacklisted. If the actor returns and triggers any ModSecurity
rule even once, they are blacklisted again based on this list.

Signed-off-by: Aaron Bedra <aaron@aaronbedra.com>