Implements analyze
This commit removes the old blacklist and ofdp options and any
additional unused options in favor of a single analyze routine. As time
goes on additional analyzers will be added to this routine. Finer
grained controls around the analyze function will eventually be made,
but isn't in scope for now.

Signed-off-by: Aaron Bedra <aaron@aaronbedra.com>